+++ lib/src/utils/logger/logger_import.dart
@@ -1,9 +1,9 @@
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-import 'package:talker/talker.dart';
 import 'dart:math' as math;
 
 import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:talker/talker.dart';
 
 import 'logs/critical_log.dart';
 import 'logs/debug_log.dart';
@@ -14,6 +14,6 @@ import 'logs/route_log.dart';
 import 'logs/verbose_log.dart';
 import 'logs/warning_log.dart';
 
-part 'logger.dart';
 part 'dio_logger.dart';
+part 'logger.dart';
 part 'observers/route_observer.dart';